+++ tests/testthat/test_get_test_index.R
@@ -1,0 +1,10 @@
+context("get_test_index")
+
+
+test_that("get_test_index works as expected", {
+
+  ti <- get_test_index()
+  ti
+
+
+})